// src/types.tsx
export type SkiddleEvent = {
  EventCode: string;
  cancellationDate: Date;
  cancellationReason: string;
  cancellationType: string;
  cancelled: string;
  countcollapsedresults: number;
  date: Date;
  description: string;
  enddate: Date;
  entryprice: string;
  eventname: string;
  eventvisibility: string;
  goingtocount: string;
  goingtos: number;
  hascollapsedresults: boolean;
  headerHex: string;
  healthAndSafety: Array<any>;
  id: string;
  imageurl: string;
  imgoing: boolean;
  largeimageurl: string;
  link: string;
  listingid: string;
  minage: string;
  openingtimes: any;
  rep: any;
  rescheduledDate: Date;
  startdate: Date;
  ticketUrl: string;
  tickets: boolean;
  uniquelistingidentifier: number;
  venue: { name: string; town: string };
  xlargeimageurl: string;
  xlargeimageurlWebP: string;
};